requirements-win.txt
PyQt6>=6.6.0
pystray>=0.19.5
Pillow>=10.2.0
pywin32>=306
pyperclip>=1.8.2
faster-whisper>=1.0.0
httpx>=0.27.0
certifi>=2024.2.2
numpy>=1.26.0
sounddevice>=0.4.6
pynput>=1.7.6
